requirements.txt
# API Development
fastapi==0.70.0
uvicorn==0.15.0
requests==2.26.0

# Database Management
psycopg2==2.9.9
pymongo==4.0.1

# Testing
pytest==7.2.2
requests-mock==1.9.3

# Containerization
docker==7.1.0
docker-compose==1.29.2

# Environment variables handling
python-dotenv